+++ test/run_all.dart
@@ -6,7 +6,6 @@ library test_run_all;
 import 'package:test/test.dart';
 
 import 'async_test.dart' as async_test;
-import 'inject_test.dart' as inject_test;
 import 'convert_test.dart' as convert_test;
 import 'util_test.dart' as util_test;
 import 'http_test.dart' as http_test;
@@ -17,6 +16,5 @@ main() {
   group("convert test", convert_test.main);
   group("util test", util_test.main);
   group("http test", http_test.main);
-  group("inject tests", inject_test.main);
   group("browser tests", browser_test.main);
 }